MANUAL: extracts mappers to their own folder
